Refactor answer request imports; rename and reorganize files for clarity
