+++ src/api/challenges/get-answer/GetAnswerRequest.ts
@@ -1,4 +1,4 @@
-import { GetAnswerResponse, GetAnswerResponseBody } from "@/api/challenges/verify-answer/GetAnswerResponse";
+import { GetAnswerResponse, GetAnswerResponseBody } from "@/api/challenges/get-answer/GetAnswerResponse";
 import { PrivilegedRequest } from "@/api/PrivilegedRequest";
 
 const METHOD = "GET";
